+++ source/Button/Button.js
@@ -3,7 +3,8 @@
  * @author vivaxy
  */
 
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import warning from 'warning';
 
 import i18n from '../config/i18n';