+++ source/HorizontalFlex/HorizontalFlexItem.js
@@ -3,7 +3,8 @@
  * @author vivaxy
  */
 
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 
 export default class HorizontalFlexItem extends Component {