+++ source/List/Item.js
@@ -3,7 +3,8 @@
  * @author vivaxy
  */
 
-import React, { Component, PropTypes, cloneElement } from 'react';
+import React, { Component, cloneElement } from 'react';
+import PropTypes from 'prop-types';
 
 import { BorderBox, BorderLine } from '../BorderBox';
 import colors from '../colors';